requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-session==0.5.0
openai==1.97.0
python-dotenv==1.0.1
httpx==0.27.0
